Tidied Main window
